--- a/.buildkite/commands/assemble-release-apk.sh
+++ b/.buildkite/commands/assemble-release-apk.sh
@@ -1,0 +1,15 @@
+#!/bin/bash -eu
+
+.buildkite/commands/restore-cache.sh
+
+echo "--- :rubygems: Setting up Gems"
+
+install_gems
+
+echo "--- :closed_lock_with_key: Installing Secrets"
+
+bundle exec fastlane run configure_apply
+
+echo "--- ⚙️ Building release variant"
+
+./gradlew assembleRelease -PskipSentryProguardMappingUpload=true