--- a/.buildkite/commands/assemble-release-apk.sh
+++ b/.buildkite/commands/assemble-release-apk.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -eu
 
-.buildkite/commands/restore-cache.sh
+"$(dirname "${BASH_SOURCE[0]}")/restore-cache.sh"
 
 echo "--- :rubygems: Setting up Gems"