--- a/.buildkite/commands/lint.sh
+++ b/.buildkite/commands/lint.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -u
 
-.buildkite/commands/restore-cache.sh
+"$(dirname "${BASH_SOURCE[0]}")/restore-cache.sh"
 
 echo "--- :rubygems: Setting up Gems"
 install_gems