--- a/.buildkite/commands/restore-cache.sh
+++ b/.buildkite/commands/restore-cache.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+set -e
+
+echo "--- 💾 Restore Cache"
+restore_gradle_dependency_cache || true