--- a/.buildkite/commands/restore-cache.sh
+++ b/.buildkite/commands/restore-cache.sh
@@ -1,6 +1,4 @@
-#!/bin/bash
-
-set -e
+#!/bin/bash -e
 
 echo "--- 💾 Restore Cache"
 restore_gradle_dependency_cache || true