--- a/.buildkite/commands/run-unit-tests.sh
+++ b/.buildkite/commands/run-unit-tests.sh
@@ -1,0 +1,7 @@
+#!/bin/bash -eu
+
+.buildkite/commands/restore-cache.sh
+
+echo "--- 🧪 Testing"
+
+./gradlew testDebugUnitTest