--- a/.buildkite/pipeline.yml
+++ b/.buildkite/pipeline.yml
@@ -33,6 +33,7 @@ steps:
 
   - label: 'Unit tests'
     command: |
+      .buildkite/commands/restore-cache.sh
       echo "--- 🧪 Testing"
       ./gradlew testDebugUnitTest
     plugins: [$CI_TOOLKIT]
@@ -52,6 +53,7 @@ steps:
 
   - label: "Instrumented tests"
     command: |
+      .buildkite/commands/restore-cache.sh
       echo "--- :rubygems: Setting up Gems"
       install_gems
       echo "--- :closed_lock_with_key: Installing Secrets"
@@ -64,6 +66,7 @@ steps:
 
   - label: "Assemble release APK"
     command: |
+      .buildkite/commands/restore-cache.sh
       echo "--- :rubygems: Setting up Gems"
       install_gems
       echo "--- :closed_lock_with_key: Installing Secrets"