Project Thread: paaHJt-7CE-p2 (Stretch Goal)
Related: [WCAndroid#133303](https://github.com/woocommerce/woocommerce-android/pull/13303) + [JP/WPAndroid#21584](https://github.com/wordpress-mobile/WordPress-Android/pull/21584)
Depends On: [a8c-ci-toolkit#135](https://github.com/Automattic/a8c-ci-toolkit-buildkite-plugin/pull/135) -> [a8c-ci-toolkit#138](https://github.com/Automattic/a8c-ci-toolkit-buildkite-plugin/pull/138)
Required By: [buildkite-ci#581](https://github.com/Automattic/buildkite-ci/pull/581)

---

## Description
<!-- Please include a summary of what this PR is changing and why these changes are needed. -->

This PR enables dependency cache on CI for this project. 💾

The dependency cache mechanism is split between:
- A weekly scheduled **save cache** build (see [commit](https://github.com/Automattic/pocket-casts-android/pull/3463/commits/7e0580591c1359b62beeb87d8c4b9595beb4ea46) and [buildkite-ci#581](https://github.com/Automattic/buildkite-ci/pull/581) PR), which targets the below tasks only:
  - Apps (`assembleDebug`)
  - Lint (`:app:lintDebug :automotive:lintDebug :wear:lintDebug`)
  - Unit Tests (`assembleDebugUnitTest`)
  - Android Tests (`assembleDebugAndroidTest`)
- A **restore cache** step that is run on targeted [pipeline.yml](https://github.com/Automattic/pocket-casts-android/blob/main/.buildkite/pipeline.yml) related jobs, via their corresponding individual scripts:
  - Mobile App (`prototype-build.sh`)
  - Lint (`lint.sh`)
  - Unit Tests (`run-unit-tests.sh`)
  - Android tests (`run-instrumented-tests.sh`)

PS: Dependency cache isn't enabled on the [release-builds.yml](https://github.com/Automattic/pocket-casts-android/blob/main/.buildkite/release-builds.yml) pipeline just yet, but just temporarily, and until we become comfortable with running this new process on the main [pipeline.yml](https://github.com/Automattic/pocket-casts-android/blob/main/.buildkite/pipeline.yml) pipeline.

---

### Future Improvement

Me and @wzieba has been discussing on DMs about some possible future improvement, which are including but are not limited to the below list:

1. Only execute the `assemble android test` task, which depends on `assemble` task anyway, and would most likely resolve most of the dependencies, for all other tasks to benefit from. However, this needs further testing to make sure that running other tasks such us `lint` and `assemble unit test` are providing little to no help.
2. Create a custom Gradle task and utilize Gradle's [init scripts](https://docs.gradle.org/current/userguide/init_scripts.html) (`init.d`). Using the init script, go through all available modules for a project, and then, using that custom Gradle task, resolve all dependencies for all tasks that are available per module. We could take some inspiration from plugins such as [this](https://plugins.gradle.org/plugin/org.gradle.github-dependency-graph-gradle-plugin).
3. Instead of scheduling the `dependency cache` build to run weekly ([buildkite-ci#581](https://github.com/Automattic/buildkite-ci/pull/581)), make it run more often, possibly scheduling it to even run daily. However, since dependencies are not being updated too often, and, even when they do on that specific week, it would only be for a few of them. This will indeed cause some extra network requests, but most of the dependencies cached would still benefit from the existing and weekly refreshed dependency cache.
4. Enable dependency cache on the [release-builds.yml](https://github.com/Automattic/pocket-casts-android/blob/main/.buildkite/release-builds.yml) pipeline.

---

## Testing Instructions
<!-- Please include step by step instructions on how to test this PR. -->
<!-- 1. Tap on the Filters tab -->
<!-- 2. Tap on a filter -->
<!-- 3. etc. -->

- `Save Cache`:

  - Triggering [schedules/dependency-cache.yml](https://github.com/Automattic/pocket-casts-android/pull/3463/files#diff-f1eb78f1f1134bf64e6d4a27fd0b816ae7dff3a0da45f4115e8898bc6939642e) manually via [this](https://buildkite.com/automattic/pocket-casts-android/builds/10175) build:
    - Starts [downloading the dependencies per task](https://buildkite.com/automattic/pocket-casts-android/builds/10175#01948435-d35c-459f-a4a9-d9ff7b2e135a/178).
    - Caches [the dependencies successfully](https://buildkite.com/automattic/pocket-casts-android/builds/10175#01948435-d35c-459f-a4a9-d9ff7b2e135a/797-798).
      - Duration: [45s](https://buildkite.com/automattic/pocket-casts-android/builds/10175#01948435-d35c-459f-a4a9-d9ff7b2e135a/797-808)
      - Size: [607.47 MB](https://buildkite.com/automattic/pocket-casts-android/builds/10175#01948435-d35c-459f-a4a9-d9ff7b2e135a/797-809)
    - Total time it takes for the `dependency cache` job to complete: [17m 9s](https://buildkite.com/automattic/pocket-casts-android/builds/10175#01948435-d35c-459f-a4a9-d9ff7b2e135a)
  - Sending a Slack notification (`#android-core-notifs`):
    - On success, notifying that: `The dependency cache has been updated.` (p1737386018597179-slack-C0787KRDYDC)
    - On failure, notifying that: `Failure to update the dependency cache.` (p1737382895136969-slack-C0787KRDYDC)
- `Restore Cache`:

| Job | BT (Without) | BT (With) | BT (Diff*) | NR (Without) | NR (With) | RC (Without) | RC (With) | Win |
| ---- | ------------ | ---------- | ---- | ------------- | --------- | -------------- | --------- | ---- |
| Assemble Release  | [6m 49s](https://buildkite.com/automattic/pocket-casts-android/builds/10171#019483fc-01bc-46c5-87e4-b4dfbf86f84a) | [5m 47s](https://buildkite.com/automattic/pocket-casts-android/builds/10177#0194846a-49a1-43dc-b2a4-a6d572274a61) | `-2m 2s` | [1m 6s](https://scans.gradle.com/s/gv6nnc5wa2f5k/performance/network-activity) | [0s](https://scans.gradle.com/s/3lm3tsbaribmq/performance/network-activity) (ish) | `0s` | [16s](https://buildkite.com/automattic/pocket-casts-android/builds/10177#0194846a-49a1-43dc-b2a4-a6d572274a61/178-187) | `50s` |
| Mobile App | [7m 5s](https://buildkite.com/automattic/pocket-casts-android/builds/10171#019483fc-01be-43a7-91cf-6f45b76961aa) | [5m 0s](https://buildkite.com/automattic/pocket-casts-android/builds/10177#0194846a-49a3-4b6f-b856-e8043946d12a) | `-2m 5s` | [1m 6s](https://scans.gradle.com/s/qzvxduq7vnadi/performance/network-activity) | [0s](https://scans.gradle.com/s/vaxr3yxeeq2uc/performance/network-activity) (ish) | `0s` | [12s](https://buildkite.com/automattic/pocket-casts-android/builds/10177#0194846a-49a3-4b6f-b856-e8043946d12a/179-188) | `54s` |
| Lint | [6m 47s](https://buildkite.com/automattic/pocket-casts-android/builds/10171#019483fc-01b5-4a59-95eb-bc0b373a8ab0) | [5m 10s](https://buildkite.com/automattic/pocket-casts-android/builds/10177#0194846a-4995-4d6e-b25f-141cacd1f615) | `-1m 37s` | [1m 1s](https://scans.gradle.com/s/gseg3dvf5c6h2/performance/network-activity) | [0s](https://scans.gradle.com/s/2gxuateqfxoak/performance/network-activity) (ish) | `0s` | [14s](https://buildkite.com/automattic/pocket-casts-android/builds/10177#0194846a-4995-4d6e-b25f-141cacd1f615/178-187) | `47s` |
| Unit Tests | [5m 45s](https://buildkite.com/automattic/pocket-casts-android/builds/10171#019483fc-01b8-4726-a4f9-f8f167bb39d0) | [4m 24s](https://buildkite.com/automattic/pocket-casts-android/builds/10177#0194846a-499b-49a5-a571-c9721ca7011e) | `-1m 21s` | [1m 6s](https://scans.gradle.com/s/sufdcfw4s2yja/performance/network-activity) | [0s](https://scans.gradle.com/s/j457ekpyqywfw/performance/network-activity) | `0s` | [14s](https://buildkite.com/automattic/pocket-casts-android/builds/10177#0194846a-499b-49a5-a571-c9721ca7011e/178-187) | `50s` |
| UI Tests | [11m 17s](https://buildkite.com/automattic/pocket-casts-android/builds/10171#019483fc-01bb-4ec0-b368-d1f6e83027e4) | [10m 56s](https://buildkite.com/automattic/pocket-casts-android/builds/10177#0194846a-499f-455d-bf4a-50893f78d991) | `-21s` | [1m 10s](https://scans.gradle.com/s/aq4gjley4tbbe/performance/network-activity) | [0s](https://scans.gradle.com/s/ynosnn3yuzmb4/performance/network-activity) | `0s` | [14s](https://buildkite.com/automattic/pocket-casts-android/builds/10177#0194846a-499f-455d-bf4a-50893f78d991/178-187) | `56s` |

> [!NOTE]
> - `Without`: Without Dependency Cache
> - `With`: With Dependency Cache
> - `BT`: Build Time
> - `NR`: Network Requests
> - `RC`: Restore Cache
> - `Win`: NR (Without) - RC (With)

> [!WARNING]
> (*): The `BT (Diff)` presented above is relative, so don't read too much into it, it might not match with the `WIN`, which is a much more solid number. But still, it is nice to have and compare those two number together. To to be clear about that, although I tried to pick the most representative builds, without and with the dependency cache mechanism in place, there are other factors that can make any build faster, or slower for that matter (like `build cache`, `firebase test lab`, etc), so again, don't read too much into it.

---

## Checklist `N/A`
 
#### I have tested any UI changes... `N/A`
<!-- If this PR does not contain UI changes, ignore these items -->